% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltfObj.R
\name{Gltf}
\alias{Gltf}
\title{R6 Class for glTF file objects}
\description{
The glTF file spec is described here: \url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html}.  This
object encapsulates most of the data from those files.
}
\examples{

## ------------------------------------------------
## Method `Gltf$print`
## ------------------------------------------------

samples <- "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/master/2.0"
gltf <- readGLB(paste0(samples, "/2CylinderEngine/glTF-Binary/2CylinderEngine.glb?raw=true"))
gltf$print(names = "meshes")
}
\section{Super class}{
\code{\link[rgl2gltf:Buffer]{rgl2gltf::Buffer}} -> \code{gltf}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scene}}{The default scene number}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Gltf$new()}}
\item \href{#method-getScene}{\code{Gltf$getScene()}}
\item \href{#method-setScene}{\code{Gltf$setScene()}}
\item \href{#method-addScene}{\code{Gltf$addScene()}}
\item \href{#method-addToScene}{\code{Gltf$addToScene()}}
\item \href{#method-defaultScene}{\code{Gltf$defaultScene()}}
\item \href{#method-getNode}{\code{Gltf$getNode()}}
\item \href{#method-setNode}{\code{Gltf$setNode()}}
\item \href{#method-addNode}{\code{Gltf$addNode()}}
\item \href{#method-addChild}{\code{Gltf$addChild()}}
\item \href{#method-getCamera}{\code{Gltf$getCamera()}}
\item \href{#method-getExtras}{\code{Gltf$getExtras()}}
\item \href{#method-setExtras}{\code{Gltf$setExtras()}}
\item \href{#method-getMesh}{\code{Gltf$getMesh()}}
\item \href{#method-addMesh}{\code{Gltf$addMesh()}}
\item \href{#method-getMaterial}{\code{Gltf$getMaterial()}}
\item \href{#method-getTexture}{\code{Gltf$getTexture()}}
\item \href{#method-getImage}{\code{Gltf$getImage()}}
\item \href{#method-addMaterial}{\code{Gltf$addMaterial()}}
\item \href{#method-addTexture}{\code{Gltf$addTexture()}}
\item \href{#method-addImage}{\code{Gltf$addImage()}}
\item \href{#method-addSampler}{\code{Gltf$addSampler()}}
\item \href{#method-getMaterialNumber}{\code{Gltf$getMaterialNumber()}}
\item \href{#method-writeVectors}{\code{Gltf$writeVectors()}}
\item \href{#method-makePrimitive}{\code{Gltf$makePrimitive()}}
\item \href{#method-getAsset}{\code{Gltf$getAsset()}}
\item \href{#method-setAsset}{\code{Gltf$setAsset()}}
\item \href{#method-getTransform}{\code{Gltf$getTransform()}}
\item \href{#method-getRglMaterial}{\code{Gltf$getRglMaterial()}}
\item \href{#method-as.list}{\code{Gltf$as.list()}}
\item \href{#method-print}{\code{Gltf$print()}}
\item \href{#method-listCount}{\code{Gltf$listCount()}}
\item \href{#method-clone}{\code{Gltf$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="addAccessor">}\href{../../rgl2gltf/html/Buffer.html#method-addAccessor}{\code{rgl2gltf::Buffer$addAccessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="addBufferView">}\href{../../rgl2gltf/html/Buffer.html#method-addBufferView}{\code{rgl2gltf::Buffer$addBufferView()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="closeBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-closeBuffer}{\code{rgl2gltf::Buffer$closeBuffer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="closeBuffers">}\href{../../rgl2gltf/html/Buffer.html#method-closeBuffers}{\code{rgl2gltf::Buffer$closeBuffers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="dataURI">}\href{../../rgl2gltf/html/Buffer.html#method-dataURI}{\code{rgl2gltf::Buffer$dataURI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="getAccessor">}\href{../../rgl2gltf/html/Buffer.html#method-getAccessor}{\code{rgl2gltf::Buffer$getAccessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="getBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-getBuffer}{\code{rgl2gltf::Buffer$getBuffer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="getBufferview">}\href{../../rgl2gltf/html/Buffer.html#method-getBufferview}{\code{rgl2gltf::Buffer$getBufferview()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="load">}\href{../../rgl2gltf/html/Buffer.html#method-load}{\code{rgl2gltf::Buffer$load()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="openBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-openBuffer}{\code{rgl2gltf::Buffer$openBuffer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="openBufferview">}\href{../../rgl2gltf/html/Buffer.html#method-openBufferview}{\code{rgl2gltf::Buffer$openBufferview()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="readAccessor">}\href{../../rgl2gltf/html/Buffer.html#method-readAccessor}{\code{rgl2gltf::Buffer$readAccessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="saveOpenBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-saveOpenBuffer}{\code{rgl2gltf::Buffer$saveOpenBuffer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="setBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-setBuffer}{\code{rgl2gltf::Buffer$setBuffer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="setBufferview">}\href{../../rgl2gltf/html/Buffer.html#method-setBufferview}{\code{rgl2gltf::Buffer$setBufferview()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rgl2gltf" data-topic="Buffer" data-id="writeBuffer">}\href{../../rgl2gltf/html/Buffer.html#method-writeBuffer}{\code{rgl2gltf::Buffer$writeBuffer()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$new(json = NULL, defaultbin = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json}}{list read from glTF file.}

\item{\code{defaultbin}}{optional external binary file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getScene"></a>}}
\if{latex}{\out{\hypertarget{method-getScene}{}}}
\subsection{Method \code{getScene()}}{
Get scene object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getScene(sc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc}}{Scene number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
scene object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-scene}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setScene"></a>}}
\if{latex}{\out{\hypertarget{method-setScene}{}}}
\subsection{Method \code{setScene()}}{
Update scene record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$setScene(sc, scene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sc}}{Which scene to update}

\item{\code{scene}}{New scene record}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addScene"></a>}}
\if{latex}{\out{\hypertarget{method-addScene}{}}}
\subsection{Method \code{addScene()}}{
Add a scene object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addScene()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
scene number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addToScene"></a>}}
\if{latex}{\out{\hypertarget{method-addToScene}{}}}
\subsection{Method \code{addToScene()}}{
Add node to scene
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addToScene(scene, node)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scene}}{Scene number to modify}

\item{\code{node}}{Node number(s) to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-defaultScene"></a>}}
\if{latex}{\out{\hypertarget{method-defaultScene}{}}}
\subsection{Method \code{defaultScene()}}{
Get default scene, creating it if necessary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$defaultScene()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
scene number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNode"></a>}}
\if{latex}{\out{\hypertarget{method-getNode}{}}}
\subsection{Method \code{getNode()}}{
Get node object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getNode(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{node number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
node object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-node}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setNode"></a>}}
\if{latex}{\out{\hypertarget{method-setNode}{}}}
\subsection{Method \code{setNode()}}{
Set node object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$setNode(n, node)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{node number.}

\item{\code{node}}{New node object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addNode"></a>}}
\if{latex}{\out{\hypertarget{method-addNode}{}}}
\subsection{Method \code{addNode()}}{
Add a node object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addNode(mesh = NULL, matrix = NULL, extras = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mesh}}{A mesh number}

\item{\code{matrix}}{A matrix transformation for the node}

\item{\code{extras}}{A list of extras, typically `rgl` objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
node number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addChild"></a>}}
\if{latex}{\out{\hypertarget{method-addChild}{}}}
\subsection{Method \code{addChild()}}{
Add node as child of another
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addChild(parent, node)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{Node number to modify}

\item{\code{node}}{Node number(s) to add as children}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCamera"></a>}}
\if{latex}{\out{\hypertarget{method-getCamera}{}}}
\subsection{Method \code{getCamera()}}{
Get camera object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getCamera(cam)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cam}}{camera number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
camera object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-camera}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExtras"></a>}}
\if{latex}{\out{\hypertarget{method-getExtras}{}}}
\subsection{Method \code{getExtras()}}{
Get top-level extras list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getExtras()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
extras list, including rgl objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setExtras"></a>}}
\if{latex}{\out{\hypertarget{method-setExtras}{}}}
\subsection{Method \code{setExtras()}}{
Set extras list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$setExtras(extras)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extras}}{New extras list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMesh"></a>}}
\if{latex}{\out{\hypertarget{method-getMesh}{}}}
\subsection{Method \code{getMesh()}}{
Get mesh object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getMesh(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{mesh number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mesh object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-mesh}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMesh"></a>}}
\if{latex}{\out{\hypertarget{method-addMesh}{}}}
\subsection{Method \code{addMesh()}}{
Add a mesh object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addMesh(primitives)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{primitives}}{A list of primitive objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mesh number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-getMaterial}{}}}
\subsection{Method \code{getMaterial()}}{
Get material object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getMaterial(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{material number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
material object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-material}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTexture"></a>}}
\if{latex}{\out{\hypertarget{method-getTexture}{}}}
\subsection{Method \code{getTexture()}}{
Get texture object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getTexture(tex)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tex}}{texture number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
texture object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-texture}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getImage"></a>}}
\if{latex}{\out{\hypertarget{method-getImage}{}}}
\subsection{Method \code{getImage()}}{
Get image object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getImage(im)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{im}}{image number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
image object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-image}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-addMaterial}{}}}
\subsection{Method \code{addMaterial()}}{
Construct and possibly add material.

This will return an existing material if possible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addMaterial(mat, defaultMaterial = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{An `rgl` material record}

\item{\code{defaultMaterial}}{Default material properties}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
material number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTexture"></a>}}
\if{latex}{\out{\hypertarget{method-addTexture}{}}}
\subsection{Method \code{addTexture()}}{
Add a texture.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addTexture(mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{An `rgl` material record}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
texture number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addImage"></a>}}
\if{latex}{\out{\hypertarget{method-addImage}{}}}
\subsection{Method \code{addImage()}}{
Add an image for a texture
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addImage(mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{An `rgl` material record}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
image number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addSampler"></a>}}
\if{latex}{\out{\hypertarget{method-addSampler}{}}}
\subsection{Method \code{addSampler()}}{
Add a sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$addSampler(mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{An `rgl` material record}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sampler number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMaterialNumber"></a>}}
\if{latex}{\out{\hypertarget{method-getMaterialNumber}{}}}
\subsection{Method \code{getMaterialNumber()}}{
Add or return a material
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getMaterialNumber(material)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{material}}{A glTF material record}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
material number
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeVectors"></a>}}
\if{latex}{\out{\hypertarget{method-writeVectors}{}}}
\subsection{Method \code{writeVectors()}}{
Write data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$writeVectors(coords, target = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coords}}{data to write, or `NULL`}

\item{\code{target}}{optional target use for data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
accessor number, or `NULL`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makePrimitive"></a>}}
\if{latex}{\out{\hypertarget{method-makePrimitive}{}}}
\subsection{Method \code{makePrimitive()}}{
Create a primitive record
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$makePrimitive(inds, mode = NULL, attributes = NULL, matnum = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inds}}{Indices of vertices}

\item{\code{mode}}{Mode of primitive}

\item{\code{attributes}}{Primitive attributes}

\item{\code{matnum}}{Material number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Primitive record, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-primitive}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAsset"></a>}}
\if{latex}{\out{\hypertarget{method-getAsset}{}}}
\subsection{Method \code{getAsset()}}{
Get asset list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getAsset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
asset object, documented here:
\url{https://www.khronos.org/registry/glTF/specs/2.0/glTF-2.0.html#reference-asset}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setAsset"></a>}}
\if{latex}{\out{\hypertarget{method-setAsset}{}}}
\subsection{Method \code{setAsset()}}{
Set asset list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$setAsset(version, generator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Version number of glTF format}

\item{\code{generator}}{Identifier of code generating it}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTransform"></a>}}
\if{latex}{\out{\hypertarget{method-getTransform}{}}}
\subsection{Method \code{getTransform()}}{
Get local transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getTransform(node, parentTransform)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{Node object}

\item{\code{parentTransform}}{Matrix transform of parent object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
4x4 matrix of local transform
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRglMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-getRglMaterial}{}}}
\subsection{Method \code{getRglMaterial()}}{
Reconstruct `rgl` material
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$getRglMaterial(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Material number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`rgl` material record
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.list"></a>}}
\if{latex}{\out{\hypertarget{method-as.list}{}}}
\subsection{Method \code{as.list()}}{
Convert to list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$as.list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List suitable for writing using JSON
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print `gltf` objects with various levels of detail.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$print(verbose = FALSE, names = FALSE, showExtras = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{logical indicator of verbose printing, or
character vector of components to print verbosely}

\item{\code{names}}{Which fields to print.}

\item{\code{showExtras}}{Logical:  show extra fields?}

\item{\code{...}}{Passed...}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{samples <- "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/master/2.0"
gltf <- readGLB(paste0(samples, "/2CylinderEngine/glTF-Binary/2CylinderEngine.glb?raw=true"))
gltf$print(names = "meshes")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listCount"></a>}}
\if{latex}{\out{\hypertarget{method-listCount}{}}}
\subsection{Method \code{listCount()}}{
Get number of items in private list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$listCount(list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{list}}{Name of list to get}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gltf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
